T2/score_calculator/requirements.txt
redis
psycopg2-binary
scikit-learn
numpy
sentence-transformers
kafka-python==2.0.2